Problem: Can't make tag-protocols for trait A just extending B, C
Explanation:

Consider you have defprotocol B with some method f and C with method g.

If you want to make a protocol A, which extends both B and C, meaning
"only things that are B and C can be A", for example, in a type class
library implementation for Elixir, you can't do it without getting a
warning.

Solution:

 - Incorporate Quinn's patch which turns __before_compile__ for protocol
   into a macro, which injects impl/1 implementation, which we list as a
   callback, turning any 0-sized protocol into 1-sized behaviour. See
   the original commit message below.
 - Test it!

* * *

Original commit message:

> Avoid compile warnings when implementing an empty protocol

Implementing a protocol creates a module that implements the
behaviour given by the protocol. A protocol that defines no
functions by extension defines no callbacks, and so won't be
recognized as a behaviour by the Erlang compiler. This means
that Elixir will log a warning when implementing such a protocol,
because behaviour_info/1 won't be defined on it, and so Elixir
also won't recognize it as a behaviour.

Erlang offers no mechanism to explicitly denote a module as
defining a behaviour, but in this case we know that every protocol
implementation must define an impl/1 function. By adding a callback
that describes this function to all protocols, the warning can be
avoided.